+++ src/components/pages/Home/Home.tsx
@@ -1,0 +1,18 @@
+// Framework Imports
+import React from 'react'
+
+// CSS Imports
+import StyledHome from './StyledHome'
+
+/**
+ * This the Home Page.
+ */
+const Home = () => {
+  return (
+    <StyledHome>
+      <p>Home Page</p>
+    </StyledHome>
+  )
+}
+
+export default Home